+++ src/matrix/mri/MRI.hpp
@@ -19,7 +19,7 @@ IntensityMap (const Matrix< std::complex <T> >& sens, bool sqroot = true) {
 	{		
 		
 #pragma omp for schedule (guided)
-		for (size_t i = 0; i < nr; i++) {
+		for (int i = 0; i < nr; i++) {
 			
 			for (size_t j = 0; j < nc; j++)
 				res[i] += real(sens(i+j*nr) * TypeTraits<std::complex<T> >::Conj(sens(i+j*nr)));
